SUCCESS! Depenceny Injection appears to be configured and the app kinda still works
