--- src/.vscode/launch.json
@@ -1,22 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Python: FastAPI",
-            "type": "python",
-            "request": "launch",
-            "module": "uvicorn",
-            "args": [
-                "server.main:app",
-                "--reload",
-                "--port",
-                "7000"
-            ],
-            "jinja": true,
-            "justMyCode": true
-        }
-    ]
-}
